Adding Sound Effects
Now items have been added that have their sounds when you move something from the inventory.

Gun sounds have been added.

It was added in the Weapons configuration, that it is now possible to put textures on them, if you don't want the texture not to be loaded, just put "load = false"